- Peter Anvin: more P4 configuration parsing - Stephen Tweedie: O_SYNC patches. Make O_SYNC/fsync/fdatasync   do the right thing. - Keith Owens: make mdule loading use the right struct module size - Boszormenyi Zoltan: get MTRR's right for the >32-bit case - Alan Cox: various random documentation etc - Dario Ballabio: EATA and u14-34f update - Ivan Kokshaysky: unbreak alpha ruffian - Richard Henderson: PCI bridge initialization on alpha - Zach Brown: correct locking in Maestro driver - Geert Uytterhoeven: more m68k updates - Andrey Savochkin: eepro100 update - Dag Brattli: irda update - Johannes Erdfelt: USB update

+++ include/asm-generic/bitops.h
@@ -9,14 +9,14 @@
  * disable interrupts while they operate.  (You have to provide inline
  * routines to cli() and sti().)
  *
- * Also note, these routines assume that you have 32 bit integers.
+ * Also note, these routines assume that you have 32 bit longs.
  * You will have to change this if you are trying to port Linux to the
  * Alpha architecture or to a Cray.  :-)
  * 
  * C language equivalents written by Theodore Ts'o, 9/26/92
  */
 
-extern __inline__ int set_bit(int nr,int * addr)
+extern __inline__ int set_bit(int nr,long * addr)
 {
 	int	mask, retval;
 
@@ -29,7 +29,7 @@ extern __inline__ int set_bit(int nr,int * addr)
 	return retval;
 }
 
-extern __inline__ int clear_bit(int nr, int * addr)
+extern __inline__ int clear_bit(int nr, long * addr)
 {
 	int	mask, retval;
 
@@ -42,7 +42,7 @@ extern __inline__ int clear_bit(int nr, int * addr)
 	return retval;
 }
 
-extern __inline__ int test_bit(int nr, int * addr)
+extern __inline__ int test_bit(int nr, long * addr)
 {
 	int	mask;
